Fix auth issue in get_instance_host
_get_instance_host was attempting to use the
admin username and password with the user's
tenant, which leads to 401 auth errors from
Keystone when attempting to determine the
instance host from nova.

This patch reads the admin tenant from the same
place as the username and password and uses that
instead. It also removes the keystone call that
was used to determine the tenant_id from the name,
since that is no longer needed.